[FLINK-5619] Consolidate AggregatingState Tests in StateBackendTestBase
